+++ NEWS.md
@@ -1,3 +1,3 @@
 # OmopSketch 0.2.2.900
 
--   account for int64 in summariseInObservation by @cecicampanile
+- Account for int64 in summariseInObservation by @cecicampanile #311