+++ NEWS.md
@@ -1,4 +1,4 @@
 # OmopSketch 0.2.2.900
 
--   Account for int64 in summariseInObservation by @cecicampanile #311
--   Add "datatable" as possible table type by @cecicampanile
+-   Account for int64 in summariseInObservation by @cecicampanile #312
+-   Add "datatable" as possible table type by @cecicampanile #314