+++ NEWS.md
@@ -1,4 +1,5 @@
 # OmopSketch 0.2.2.900
 
+
 -   Account for int64 in summariseInObservation by @cecicampanile #312
 -   Add "datatable" as possible table type by @cecicampanile #314