+++ NEWS.md
@@ -1,4 +1,8 @@
 # OmopSketch 0.2.2.900
 
 -   Account for int64 in summariseInObservation by @cecicampanile #311
+
+-   Interval argument in summariseMissingData and summariseConceptIdCounts, year argument deprecated by @cecicampanile #317
+
 -   Only records in observation are accounted in summariseConceptIdCounts and summariseConceptSetCounts by @cecicampanile #319
+