+++ NEWS.md
@@ -2,6 +2,9 @@
 
 -   Account for int64 in summariseInObservation by @cecicampanile #311
 
--   eunomiaIsAvailable instead of the deprecated eunomia_is_available #316
+-   eunomiaIsAvailable instead of the deprecated eunomia_is_available by @cecicampanile #316
+
+-   Interval argument in summariseMissingData and summariseConceptIdCounts, year argument deprecated by @cecicampanile #317
 
 -   Only records in observation are accounted in summariseConceptIdCounts and summariseConceptSetCounts by @cecicampanile #319
+